Removed double cast in stopsByBbox field
The double cast for latitude and longitude values caused an error as
pointed out in issue #2510.

+++ src/main/java/org/opentripplanner/index/IndexGraphQLSchema.java
@@ -893,10 +893,8 @@ public class IndexGraphQLSchema {
                     .build())
                 .dataFetcher(environment -> index.graph.streetIndex
                     .getTransitStopForEnvelope(new Envelope(
-                        new Coordinate((double) (float) environment.getArgument("minLon"),
-                            (double) (float) environment.getArgument("minLat")),
-                        new Coordinate((double) (float) environment.getArgument("maxLon"),
-                            (double) (float) environment.getArgument("maxLat"))))
+                        new Coordinate(environment.getArgument("minLon"), environment.getArgument("minLat")),
+                        new Coordinate(environment.getArgument("maxLon"), environment.getArgument("maxLat"))))
                     .stream()
                     .map(TransitVertex::getStop)
                     .filter(stop -> environment.getArgument("agency") == null || stop.getId()
